Minor fixes
Benchedmarked the script and found it to be 10 times more efficient than
"heisenberg_chain_H.py" at small sizes but at least 10 times slower at bigger
sizes. Possibly due to the abundance of for loops in the script but I think
those are quite unavoidable. To be practical, this might need to be implemented
in Fortran/f2py.