Added some functions and applied some fixes
Updated the spin2z function to give it a significant performance boost.
Added a new function "recast" to take a state that only contains element
of a certain total Sz and augment it to include all other elements.
Rewrite energy_density to use a function from quantum_module.
Added a function to find a state with only the center block (zero total
Sz) of a Hamiltonian.