This module contains a class that does a better job at estimating time.
The EstimateTime class uses queries four algorithms and choose the best
one to make an estimate of how much longer it would take the computer to
finish all the jobs. As of now, it is giving mostly useful estimates but
it occasionally acts up. The algorithm for choosing the best algorithm
needs to be polished.